Reorganize system generator, Add logic for generating sector and storing in redux, add redux system view
