+++ src/utils/system-generator/planets.ts
@@ -1,0 +1,45 @@
+export const planetList: string[] = [
+  'Rocky1',
+  'Rocky2',
+  'Rocky3',
+  'Rocky4',
+  'Rocky5',
+  'Temperate1',
+  'Temperate2',
+  'Temperate3',
+  'Temperate4',
+  'Temperate5',
+  'Ocean1',
+  'Ocean2',
+  'Ocean3',
+  'Ocean4',
+  'Ocean5',
+  'Frozen1',
+  'Frozen2',
+  'Frozen3',
+  'Frozen4',
+  'Frozen5',
+  'Lava1',
+  'Lava2',
+  'Lava3',
+  'Lava4',
+  'Lava5',
+  'Gas1',
+  'Gas2',
+  'Gas3',
+  'Gas4',
+  'Gas5',
+  'Desert1',
+  'Desert2',
+  'Desert3',
+  'Desert4',
+  'Desert5',
+  'Greenhouse1',
+  'Greenhouse2',
+  'Greenhouse3',
+  'Greenhouse4',
+  'Greenhouse5',
+  'Asteroid-Belt1',
+  'Asteroid-Belt2',
+  'Asteroid-Belt3',
+];
