+++ src/utils/system-generator/resources.ts
@@ -1,0 +1,17 @@
+export const resources: string[] = [
+  'ore-low',
+  'ore-high',
+  'ore-medium',
+  'ore-trace',
+  'water-high',
+  'water-medium',
+  'water-low',
+  'water-trace',
+  'gas-low',
+  'gas-medium',
+  'population-tribal',
+  'population-scarce',
+  'population-cities',
+  'oil-low',
+  'oil-medium',
+];
