+++ src/utils/system-generator/stars.ts
@@ -1,0 +1,9 @@
+export const starList: string[] = [
+  'Red-Giant',
+  'Red-Supergiant',
+  'Blue-Giant',
+  'White-Dwarf',
+  'Yellow-Dwarf',
+  'Red-Dwarf',
+  'Brown-Dwarf',
+];
